reorganized the code
