+++ pkg/clients/snowflake/identifiers_test.go
